[Silver III] Title: 주유소, Time: 328 ms, Memory: 36136 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -9,19 +9,19 @@ public class Main {
 		
 		int N = Integer.parseInt(br.readLine()); //도시 개수
 		
-		int[] len = new int[N-1]; //도로의 길이
+		long[] len = new long[N-1]; //도로의 길이
 		st = new StringTokenizer(br.readLine());
 		for(int i=0; i<N-1; i++) {
 			len[i] = Integer.parseInt(st.nextToken());
 		}
 		
-		int[] price = new int[N]; //리터당 가격
+		long[] price = new long[N]; //리터당 가격
 		st = new StringTokenizer(br.readLine());
 		for(int i=0; i<N; i++) {
 			price[i] = Integer.parseInt(st.nextToken());
 		}
 		
-		int min = price[0]; //최소 비용 초기화
+		long min = price[0]; //최소 비용 초기화
 		long result = 0; //총 최소 비용
 		
 		for(int i=0; i<N-1; i++) { //최소 비용 계산
